refactor(infrastructure): merge duplicate AggregateRepository files
Replaced broken save_events/plain-list get_events calls with the
correct append_events/Result-based get_events API, narrowed
generics from Any to UUID, and removed the unused _new variant.
