+++ src/forging_blocks/infrastructure/repositories/aggregate_repository.py
@@ -1,18 +1,17 @@
-# pyright: reportAttributeAccessIssue=false, reportUnknownMemberType=false, reportArgumentType=false
 """Aggregate RepositoryPort implementation.
 
 Provides a repository specifically designed for AggregateRoot persistence
 with event sourcing support.
 """
 
-from typing import Any
+from uuid import UUID
 
 from forging_blocks.domain.aggregate_root.aggregate_root import AggregateRoot
 from forging_blocks.infrastructure.event_stores.event_store_base import EventStoreBase
 from forging_blocks.infrastructure.repositories.base_repository import BaseRepository
 
 
-class AggregateRepository[TAggregateRoot: AggregateRoot[Any, Any], TId: Any](
+class AggregateRepository[TAggregateRoot: AggregateRoot[object, UUID], TId: UUID](
     BaseRepository[TAggregateRoot, TId]
 ):
     """RepositoryPort for AggregateRoot persistence with event sourcing.
@@ -20,6 +19,8 @@ class AggregateRepository[TAggregateRoot: AggregateRoot[Any, Any], TId: Any](
     Extends BaseRepository with event store integration for
     event-sourced aggregates.
     """
+
+    _event_store: EventStoreBase[object]
 
     def __init__(
         self,
@@ -41,17 +42,12 @@ class AggregateRepository[TAggregateRoot: AggregateRoot[Any, Any], TId: Any](
         Args:
             aggregate: The aggregate to save.
         """
-        # Save the aggregate snapshot
         await super().save(aggregate)
-
-        # Persist uncommitted events to the event store
         events = aggregate.collect_events()
-        if events:
-            await self._event_store.save_events(
-                aggregate_id=str(aggregate.id),
-                events=[event.to_dict() for event in events],
-                expected_version=aggregate.version.value - len(events),
-            )
+        aggregate_id: UUID | None = aggregate.id  # type: ignore[assignment]
+        if events and aggregate_id is not None:
+            version = aggregate.version.value - len(events)
+            await self._event_store.append_events(aggregate_id, events, expected_version=version)
 
     async def get_by_id(self, id: TId) -> TAggregateRoot | None:  # noqa: A002
         """Retrieve an aggregate by ID and replay its events.
@@ -62,17 +58,10 @@ class AggregateRepository[TAggregateRoot: AggregateRoot[Any, Any], TId: Any](
         Returns:
             The retrieved aggregate or None if not found.
         """
-        # First check in-memory storage
         aggregate = await super().get_by_id(id)
         if aggregate is not None:
             return aggregate
-
-        # If not in memory, try to load from event store
-        events = await self._event_store.get_events(str(id))
-        if not events:
+        result = await self._event_store.get_events(id)  # type: ignore[arg-type]
+        if not result.is_ok or not result.value:
             return None
-
-        # Reconstruct aggregate from events
-        # This requires the aggregate type to have a factory method
-        # For now, return None - concrete implementations should override
         return None
